+++ src/app/(default)/book/[id]/page.tsx
@@ -12,7 +12,7 @@ import {
 import { DEFAULT_LIMIT } from '@/constants/api';
 import { GetBook, SearchReview } from '@frolog/frolog-api';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/utils/auth/auth';
+import { authOptions } from '@/utils/auth/nextAuth';
 import { QUERY_KEY } from '@/constants/query';
 
 interface Props {
