+++ src/app/(user)/[userId]/well/[wellId]/book/[bookId]/(list)/review/page.tsx
@@ -4,7 +4,7 @@ import dynamic from 'next/dynamic';
 import { getIsRootUser } from '@/utils/auth/getIsRootUser';
 import { Metadata } from 'next';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/utils/auth/auth';
+import { authOptions } from '@/utils/auth/nextAuth';
 import {
   dehydrate,
   HydrationBoundary,
