+++ src/app/(user)/[userId]/well/create/page.tsx
@@ -1,6 +1,6 @@
 import { QUERY_KEY } from '@/constants/query';
 import { WellForm } from '@/features/Well';
-import { authOptions } from '@/utils/auth/auth';
+import { authOptions } from '@/utils/auth/nextAuth';
 import { SearchStoreItem } from '@frolog/frolog-api';
 import {
   dehydrate,
