+++ src/utils/auth/getIsRootUser.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from 'next-auth';
-import { authOptions } from './auth';
+import { authOptions } from './nextAuth';
 
 /** 현재 로그인된 유저와 같은 유저인지 확인하는 함수 */
 export const getIsRootUser = async (userId: string) => {
